refactor: move dock context menu implementation to loader module
1. Deleted old dock context menu files from plugins/dde-dock/common
2. Created new simplified DockContextMenu class in src/loader
3. Simplified menu implementation by removing width calculation and
Wayland specific code
4. Changed PluginItem to use new DockContextMenu class
5. Moved reminder dot painting logic directly into paintEvent
6. Updated CMakeLists.txt to include new files

The change was made to:
1. Reduce complexity by removing unused features
2. Move menu implementation closer to where it's actually used
3. Simplify maintenance by having a single implementation
4. Remove dependency on Wayland-specific code in common module

refactor: 将停靠栏上下文菜单实现移至加载器模块

1. 从 plugins/dde-dock/common 删除旧的停靠栏上下文菜单文件
2. 在 src/loader 中创建新的简化版 DockContextMenu 类
3. 通过移除宽度计算和Wayland特定代码简化菜单实现
4. 修改 PluginItem 使用新的 DockContextMenu 类
5. 将提醒点绘制逻辑直接移到 paintEvent 中
6. 更新 CMakeLists.txt 包含新文件

这些修改的目的：
1. 通过移除未使用功能降低复杂性
2. 将菜单实现移到实际使用的位置
3. 通过单一实现简化维护
4. 移除公共模块中对Wayland特定代码的依赖

pms: TASK-377199
